Move GrTags.hpp and GrTagsDeclarations.hpp
